+++ hata/discord/message/attachment/__init__.py
@@ -1,10 +1,12 @@
 from .attachment import *
 from .constants import *
 from .fields import *
+from .flags import *
 
 
 __all__ = (
     *attachment.__all__,
     *constants.__all__,
     *fields.__all__,
+    *flags.__all__,
 )