+++ hata/discord/message/attachment/tests/test__Attachment__utility.py
@@ -1,6 +1,7 @@
 import vampytest
 
 from ..attachment import Attachment
+from ..flags import AttachmentFlag
 
 from .test__Attachment__constructor import _assert_fields_set
 
@@ -14,6 +15,7 @@ def test__Attachment__copy__0():
     content_type = 'application/json'
     description = 'Nue'
     duration = 12.6
+    flags = AttachmentFlag(12)
     height = 1000
     name = 'i miss you'
     size = 999
@@ -26,6 +28,7 @@ def test__Attachment__copy__0():
         content_type = content_type,
         description = description,
         duration = duration,
+        flags = flags,
         height = height,
         name = name,
         size = size,
@@ -74,6 +77,7 @@ def test__Attachment__copy_with__0():
     content_type = 'application/json'
     description = 'Nue'
     duration = 12.6
+    flags = AttachmentFlag(12)
     height = 1000
     name = 'i miss you'
     size = 999
@@ -86,6 +90,7 @@ def test__Attachment__copy_with__0():
         content_type = content_type,
         description = description,
         duration = duration,
+        flags = flags,
         height = height,
         name = name,
         size = size,
@@ -134,6 +139,7 @@ def test__Attachment__copy_with__2():
     old_content_type = 'application/json'
     old_description = 'Nue'
     old_duration = 12.6
+    old_flags = AttachmentFlag(12)
     old_height = 1000
     old_name = 'i miss you'
     old_size = 999
@@ -145,6 +151,7 @@ def test__Attachment__copy_with__2():
     new_content_type = 'image/png'
     new_description = 'Remilia'
     new_duration = 69.4
+    new_flags = AttachmentFlag(3)
     new_height = 702
     new_name = 'Slave of Scarlet'
     new_size = 701
@@ -157,6 +164,7 @@ def test__Attachment__copy_with__2():
         content_type = old_content_type,
         description = old_description,
         duration = old_duration,
+        flags = old_flags,
         height = old_height,
         name = old_name,
         size = old_size,
@@ -170,6 +178,7 @@ def test__Attachment__copy_with__2():
         content_type = new_content_type,
         description = new_description,
         duration = new_duration,
+        flags = new_flags,
         height = new_height,
         name = new_name,
         size = new_size,
@@ -185,6 +194,7 @@ def test__Attachment__copy_with__2():
     vampytest.assert_eq(copy.content_type, new_content_type)
     vampytest.assert_eq(copy.description, new_description)
     vampytest.assert_eq(copy.duration, new_duration)
+    vampytest.assert_eq(copy.flags, new_flags)
     vampytest.assert_eq(copy.height, new_height)
     vampytest.assert_eq(copy.name, new_name)
     vampytest.assert_eq(copy.size, new_size)